refactor: 내부 모듈에서 상위 배럴 재사용하지 않도록 수정(코드래빗 리뷰 반영)
